moves xml_collector into libs/templating
