+++ openvair/libs/templating/xml_collector.py
